refactor(tests): move helpers out of specs
